+++ src/binary/audio/synth/oscillators/hq/phasors/sample_organ.h
@@ -1,6 +1,6 @@
 #ifndef SAMPLE_ORGAN
 #define SAMPLE_ORGAN
 
-int sample_organ(int freq, int t);
+double sample_organ(int freq, int t);
 
 #endif
