+++ src/binary/audio/synth/oscillators/hq/phasors/sample_sawtooth.h
@@ -1,6 +1,6 @@
 #ifndef SAMPLE_SAWTOOTH
 #define SAMPLE_SAWTOOTH
 
-int sample_sawtooth(int freq, int t);
+double sample_sawtooth(int freq, int t);
 
 #endif
