+++ src/binary/audio/synth/oscillators/hq/phasors/sample_triangle.h
@@ -1,6 +1,6 @@
 #ifndef SAMPLE_TRIANGLE
 #define SAMPLE_TRIANGLE
 
-int sample_triangle(int freq, int t);
+double sample_triangle(int freq, int t);
 
 #endif
